Extract PublishablePostEditor to its own file
